+++ src/tc/azext_tc/_client_factory.py
@@ -2,6 +2,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
+# pylint: disable=import-outside-toplevel
 
 from azure.cli.core.commands.client_factory import get_mgmt_service_client