+++ src/tc/HISTORY.rst
@@ -5,12 +5,21 @@ Release History
 
 0.3.0
 ++++++
-* Preview release.
+* Preview release
+* Fix version validator
+* Fix missing help for several parameters
 * Update metadata from Alpha to Beta
 * Change max core CLI version to 3.0.0
 * Drop support for python 3.5
+* tc create -> tc deploy
 * Simplify tc provider deploy parameters
-* Return object instead of strings for tc create/upgrade, tc provider deploy/upgrade
+* Return object instead of strings deploy/upgrade
+* Updated teamcloud python client
+* Added some options to allow redeployment using tc deploy
+* Create system Managed Identitiy for function apps
+* Allow project type names without a period
+* Removed 3+ subscriptions requirement for project types
+* Fix project type create provider validation
 
 0.2.2
 ++++++